+++ changelog.md
@@ -7,6 +7,7 @@
 - Fix crash when pressing `=` in file preview
 ##### Misc
 - Change how to specify custom keybindings
+- 'yank_to_clipboard' now belongs to mode 'global' (before: 'normal')
 
 ## v0.16.1: July 03, 2018
 ##### Features